add utils.py, order the input param
