+++ mmyolo/models/task_modules/assigners/__init__.py
@@ -1,4 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .batch_atss_assigner import BatchATSSAssigner
 from .batch_task_aligned_assigner import BatchTaskAlignedAssigner
-__all__ = ['BatchATSSAssigner', 'BatchTaskAlignedAssigner']
+from .utils import select_candidates_in_gts, select_highest_overlaps
+
+__all__ = ['BatchATSSAssigner', 'BatchTaskAlignedAssigner',
+           'select_candidates_in_gts', 'select_highest_overlaps']
